Lecture 7: Change working with default state after clicking on buttons
The goal was to change the state to the default value as soon as
possible, in order to make React show the Loader component.

However, it might not be immediately clear why the previous version
works. When updating a student, it's especially difficult to explain, so
it's better to remove it altogether. Within StudentList, it is hopefully
slightly more comprehensible when it is done immediately before
re-fetching the student list.